Bug 1820467 - PR change. Move build check to caller. Updated dependencies for Focus and Sample Browser
