+++ android-components/components/support/ktx/src/main/res/values/strings.xml
@@ -13,5 +13,5 @@
     <string name="mozac_support_ktx_menu_share_with">Share with…</string>
     <string name="mozac_support_ktx_share_dialog_title">Share via</string>
     <!-- Text for confirmation "snackbar" shown after copying an image to the clipboard. -->
-    <string name="mozac_support_ktx_snackbar_delegate_image_copied" moz:removedIn="112" tools:ignore="UnusedResources">Image copied to clipboard</string>
+    <string name="mozac_support_ktx_snackbar_delegate_image_copied" moz:removedIn="114" tools:ignore="UnusedResources">Image copied to clipboard</string>
 </resources>
