+++ android-components/samples/browser/build.gradle
@@ -120,6 +120,7 @@ dependencies {
 
     implementation project(':ui-autocomplete')
     implementation project(':ui-tabcounter')
+    implementation project(':ui-widgets')
 
     // Add a dependency on service-glean to simplify the testing workflow
     // for engineers that want to test Gecko metrics exfiltrated via the Glean
