+++ fenix/app/src/main/java/org/mozilla/fenix/browser/BaseBrowserFragment.kt
@@ -499,11 +499,7 @@ abstract class BaseBrowserFragment :
             store = store,
             tabId = customTabSessionId,
             onCopyConfirmation = {
-                FenixSnackbarDelegate(binding.dynamicSnackbarContainer).show(
-                    snackBarParentView = binding.dynamicSnackbarContainer,
-                    text = R.string.snackbar_copy_image_to_clipboard_confirmation,
-                    duration = Snackbar.LENGTH_LONG,
-                )
+                showSnackbarForClipboardCopy()
             },
         )
 
@@ -906,6 +902,22 @@ abstract class BaseBrowserFragment :
         )
 
         initializeEngineView(toolbarHeight)
+    }
+
+    /**
+     * Show a [Snackbar] when data is set to the device clipboard. To avoid duplicate displays of
+     * information only show a [Snackbar] for Android 12 and lower.
+     *
+     * [See details](https://developer.android.com/develop/ui/views/touch-and-input/copy-paste#duplicate-notifications).
+     */
+    private fun showSnackbarForClipboardCopy() {
+        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.S_V2) {
+            FenixSnackbarDelegate(binding.dynamicSnackbarContainer).show(
+                snackBarParentView = binding.dynamicSnackbarContainer,
+                text = R.string.snackbar_copy_image_to_clipboard_confirmation,
+                duration = Snackbar.LENGTH_LONG,
+            )
+        }
     }
 
     /**
