+++ focus-android/app/build.gradle
@@ -287,6 +287,7 @@ dependencies {
     implementation project(':ui-colors')
     implementation project(':ui-icons')
     implementation project(':ui-tabcounter')
+    implementation project(':ui-widgets')
     implementation project(':feature-webcompat')
     implementation project(':feature-webcompat-reporter')
     implementation project(':support-webextensions')
